Make use of new safe-stable-stringify 2.x features (#134)
This commit originally included an upgrade of the lib itself from 1.x to 2.x.
After a Dependabot PR was merged in to do the upgrade, this commit was rebased to only include the changes that take advantage of the new features.
https://github.com/BridgeAR/safe-stable-stringify/releases

----

- Allow `opts` to configure the lib.
- Update types for new opts.
- Add tests to fill in coverage for json.js.

NB. The `Buffer.toString('base64')` conversion was removed. Buffer implements `toJSON` so this block of code was not being executed anyway, as the replacer is called with the result of `toJSON`. Base64 was never being returned.
https://nodejs.org/api/buffer.html#buftojson